user auth should work now hihi

+++ client/emerald-app/netlify/functions/delete_script/delete_script.js
@@ -1,8 +1,12 @@
 const { connectToDatabase } = require("../mongoDB")
 const { ObjectId } = require("mongodb")
+const { verify_jwt } = require("../verify_token")
 
 const handler = async (event) => {
   try {
+
+    const decodedUser = verify_jwt(event.headers)
+
     const requestData = JSON.parse(event.body)
     const database = await connectToDatabase()
     const collection = database.collection(process.env.MONGODB_COLLECTION_SCRIPTS);
@@ -25,6 +29,13 @@ const handler = async (event) => {
             statusCode: 400,
             body: "No script found with given ID."
         }
+    }
+
+    if(!(found_script.owner === decodedUser.user_id))
+    {
+      return{
+        statusCode: 401
+      }
     }
     
     await collection.deleteOne({
